build the sequence index
This is an index that's meant to be primarily disk-backed, as sequence vector
serialized to a file with a CSA and rank/select dictionary indexing the names
and pointing into this file. The idea is that we don't want to load the whole
sequence set into memory. We can hopefully hold all the sequence names in
memory though.