pad blindly to sort less
To avoid a two-step sort, use the max key information provided to
dmultimap::index to pad a key/nullvalue pair for every key in the key range.

This can be done prior to sorting, saving a second sorting pass after padding.

The one drawback is that we always must check to see if the value we get is
null, and skip if it is. Maybe we should be doing this anyway.

This seems to be about 10% faster for small data sets (chrI for 9 yeast genomes).