remove apparent race, segfault trigger
It's not clear to me why this is causing a form of race condition or segfault.
Perhaps it has to do with the dynamic scheduling. In any case, it's not
necessary to run this loop in parallel.

I was able to reproduce segfaults by pushing the thread count very high (100)
on my laptop. On a big SSD server with 20 CPUs this was triggered at -t 20 and
-t 8. I'm moving to test there to see if this resolves the same problem.