move back to cgranges in expectation of atomic bitvector
The idea is to record which ranges we've seen during transclosure using
an atomic bitvector over the set of ranges, and to do the exploration in
parallel. This is still a few steps out, but here we get ready for it,
and at worst we might save memory relative to using a std::set here.