output repeat-max elided bases to improve runtime
The last commit has us not writing repeat-max'ed bases until the next
loop through the transitive closure processing. This is really
inefficient, as we have to keep exploring the links in the implicit
interval tree to re-find them. This is fixed here by just putting them
in a list of bases to emit later.